refactor: toXCommand to parser private method
